requirements.txt
llama-index-core>=0.10.0
llama-index-readers-file
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-vector-stores-chroma
arxiv
PyMuPDF
python-dotenv
chromadb
tqdm